+++ jeap-error-handling-service/src/main/java/ch/admin/bit/jeap/errorhandling/infrastructure/kafka/FatalEhsProcessingException.java
@@ -1,0 +1,7 @@
+package ch.admin.bit.jeap.errorhandling.infrastructure.kafka;
+
+public class FatalEhsProcessingException extends RuntimeException {
+    public FatalEhsProcessingException(Throwable cause) {
+        super("Classified 'fatal': " + cause.getMessage(), cause);
+    }
+}