+++ jeap-error-handling-service/src/main/java/ch/admin/bit/jeap/errorhandling/infrastructure/kafka/RecoverableEhsProcessingException.java
@@ -1,0 +1,7 @@
+package ch.admin.bit.jeap.errorhandling.infrastructure.kafka;
+
+public class RecoverableEhsProcessingException extends RuntimeException {
+    public RecoverableEhsProcessingException(Throwable cause) {
+        super("Classified 'recoverable': " + cause.getMessage(), cause);
+    }
+}