+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [12.2.0] - 2025-10-07
+
+### Changed
+
+- The error handling service will now retry the processing of failed messages when temporary problems occur
+  (e.g., database connection issues) and will no longer send the messages to the dead letter topic in this case.
+
 
 ## [12.1.3] - 2025-10-07